# docs/swagger.yaml
definitions:
...
  dto.LoginResp:
    properties:
      token:
        type: string
    type: object
  dto.Resp:
    properties:
      cate_id:
        type: integer
      code:
        type: integer
      cur_page:
        type: integer
      data:
        type: object
      msg:
        type: string
      total:
        type: integer
    type: object
  dto.UserRequest:
    properties:
      code:
        type: string
      passwd:
        type: string
      phone:
        type: integer
    type: object
...
  go_micro_service_user.UserInfo:
    properties:
      phone:
        type: integer
      user_id:
        type: integer
    type: object
info:
  contact:
    email: scg130@163.com
    name: scg130
  description: micro
  title: micro
  version: "1.0"
paths:
...
  /find/{phone}:
    get:
      description: 通过手机号查找用户
      parameters:
      - description: 手机号
        in: path
        name: phone
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Resp'
            - properties:
                data:
                  $ref: '#/definitions/go_micro_service_user.UserInfo'
              type: object
        "500":
          description: 服务异常
          schema:
            type: string
      summary: 查找用户
      tags:
      - 用户中心
  /login:
    post:
      description: 登录
      parameters:
      - description: body参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Resp'
            - properties:
                data:
                  $ref: '#/definitions/dto.LoginResp'
              type: object
      summary: 登录
      tags:
      - 用户中心
...
  /register:
    post:
      description: 注册
      parameters:
      - description: body参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Resp'
        "500":
          description: 服务异常
          schema:
            type: string
      summary: 注册
      tags:
      - 用户中心
schemes:
- http
swagger: "2.0"